requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0

# Security dependencies
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
bcrypt>=4.1.0
scapy>=2.4.5
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
joblib>=1.0.1

# Machine Learning
imbalanced-learn>=0.8.0
scipy>=1.7.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-asyncio>=0.15.1
httpx>=0.18.2

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.3.6

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Monitoring
prometheus-client>=0.12.0
psutil>=5.9.0

# MongoDB
pymongo>=4.0.0
motor>=2.5.0  # Async MongoDB driver

# Configuration Management
pydantic[dotenv]>=1.8.2
